Remove unsafe changes

+++ source/virtualBuffers/__init__.py
@@ -631,7 +631,7 @@ class VirtualBuffer(browseMode.BrowseModeDocumentTreeInterceptor):
 			try:
 				watchdog.cancellableExecute(
 					NVDAHelper.localLib.VBuf_destroyBuffer,
-					ctypes.byref(self.VBufHandle),
+					ctypes.byref(ctypes.c_int(self.VBufHandle)),
 				)
 			except WindowsError:
 				pass
